refactor: move email and smtp config ports into ferriskey-domain
