+++ core/src/domain/common/email.rs
@@ -1,13 +1,4 @@
-use crate::domain::{common::entities::app_errors::CoreError, realm::entities::SmtpConfig};
-
-#[cfg_attr(test, mockall::automock)]
-pub trait EmailPort: Send + Sync {
-    fn send_email(
-        &self,
-        config: &SmtpConfig,
-        to_email: &str,
-        subject: &str,
-        body: &str,
-        html_body: Option<String>,
-    ) -> impl Future<Output = Result<(), CoreError>> + Send;
-}
+//! The `EmailPort` trait now lives in the shared `ferriskey-domain` crate (it only references
+//! `CoreError` and `realm::SmtpConfig`, both now in `ferriskey-domain`). Re-exported so existing
+//! `crate::domain::common::email::{EmailPort, MockEmailPort}` call sites keep resolving.
+pub use ferriskey_domain::common::email::*;
