+++ core/src/domain/realm/entities.rs
@@ -1,56 +1,10 @@
-use chrono::{DateTime, Utc};
-pub use ferriskey_domain::realm::{LoginAlias, LoginAliases, Realm, RealmId, RealmSetting};
+pub use ferriskey_domain::realm::{
+    LoginAlias, LoginAliases, Realm, RealmId, RealmSetting, SmtpConfig, SmtpEncryption,
+};
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
-use uuid::Uuid;
 
 use crate::domain::abyss::identity_provider::entities::IdentityProviderPresentation;
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
-pub struct SmtpConfig {
-    pub id: Uuid,
-    pub realm_id: Uuid,
-    pub host: String,
-    pub port: u16,
-    pub username: String,
-    #[serde(skip_serializing)]
-    pub password: String,
-    pub from_email: String,
-    pub from_name: String,
-    pub encryption: SmtpEncryption,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum SmtpEncryption {
-    Tls,
-    StartTls,
-    None,
-}
-
-impl SmtpEncryption {
-    pub fn as_str(&self) -> &str {
-        match self {
-            SmtpEncryption::Tls => "tls",
-            SmtpEncryption::StartTls => "starttls",
-            SmtpEncryption::None => "none",
-        }
-    }
-}
-
-impl std::str::FromStr for SmtpEncryption {
-    type Err = std::convert::Infallible;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "starttls" => SmtpEncryption::StartTls,
-            "none" => SmtpEncryption::None,
-            _ => SmtpEncryption::Tls,
-        })
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ToSchema)]
 pub struct RealmLoginSetting {
