+++ libs/ferriskey-domain/src/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod app_errors;
+pub mod email;
 pub mod policies;
